got rid of ref_gene_id
It has been replaced with gene_name, and queries suitably simplified.
No query references this column anymore, it is no longer included in
result dataframes.  Where immediately adjacent and appropriate, gene_id
has also been replaced.

This patch is guaranteed to change some of the results obtaineed from
the app.  In fact, a surprising omission was the motivation to go
forward with it.  I'd argue that results are now strictly better.