switches to external storage of expression values
Expression and coverage values are now stored in an auxilliary file, see
docs of stringtie-db for details on how it is done.  The storage is much
more compact, and most probably faster than using SQLite tables.

A tiny extension module is needed, which is now included.  To compile
the extension, simply call 'make'.